# back/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: AmDm Guitar Chords API
  version: 0.2.0

servers:
  - url: /api
    description: API base

paths:
  /artists:
    get:
      operationId: listArtists
      summary: List artists
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistList'
    post:
      operationId: createArtist
      summary: Create artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistRequest'
      responses:
        '201':
          description: Created artist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Validation error

  /artists/{artistSlug}:
    get:
      operationId: getArtistBySlug
      summary: Get artist by slug
      parameters:
        - name: artistSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artist with songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistWithSongs'
        '404':
          description: Not found

  /songs:
    get:
      operationId: listSongs
      summary: List songs
      parameters:
        - name: artist_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongList'
    post:
      operationId: createSong
      summary: Create song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSongRequest'
      responses:
        '201':
          description: Created song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '400':
          description: Validation error

  /search:
    get:
      operationId: search
      summary: Search songs by artist name/slug or song title
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of matching songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongList'

  /songs/{songId}:
    get:
      operationId: getSong
      summary: Get song by ID
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '404':
          description: Not found
    put:
      operationId: updateSong
      summary: Update song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSongRequest'
      responses:
        '200':
          description: Updated song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '400':
          description: Validation error
        '404':
          description: Not found

  /songs/{songId}/transpose:
    post:
      operationId: transposeSong
      summary: Transpose song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: semitones
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transposed song content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '404':
          description: Not found

components:
  schemas:
    Artist:
      type: object
      properties:
        artist_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - artist_id
        - name
        - slug

    ArtistWithSongs:
      allOf:
        - $ref: '#/components/schemas/Artist'
        - type: object
          properties:
            songs:
              type: array
              items:
                $ref: '#/components/schemas/SongListItem'

    ArtistList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        total:
          type: integer

    CreateArtistRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
        slug:
          type: string

    SongListItem:
      type: object
      properties:
        song_id:
          type: string
          format: uuid
        artist_id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        tonality:
          type: integer
      required:
        - song_id
        - title
        - slug

    Song:
      type: object
      properties:
        song_id:
          type: string
          format: uuid
        artist_id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        tonality:
          type: integer
        content:
          $ref: '#/components/schemas/TabContent'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - song_id
        - title
        - slug

    SongList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SongListItem'
        total:
          type: integer

    TabContent:
      type: object
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/Section'
        ascii_tab:
          type: string

    Section:
      type: object
      properties:
        type:
          type: string
        label:
          type: string
        chord_sequence:
          type: array
          items:
            type: string
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/Block'

    ChordSegment:
      type: object
      required:
        - chord
      properties:
        chord:
          type: string
        text:
          type: string
          description: "Lyrics fragment; empty string allowed (chord with no words)"

    Block:
      type: object
      required:
        - kind
      properties:
        kind:
          type: string
          enum:
            - instrumental
            - lyrics
        label:
          type: string
          description: "Optional label for instrumental blocks (e.g. Intro, Proigrysh)"
        chords:
          type: array
          items:
            type: string
          description: "Only for kind=instrumental; sequence of chords"
        tab:
          type: string
          description: "Only for kind=instrumental; optional ASCII tab snippet"
        segments:
          type: array
          items:
            $ref: '#/components/schemas/ChordSegment'
          description: "Only for kind=lyrics; chord at segment start = switch point; text may be empty"

    CreateSongRequest:
      type: object
      required:
        - artist_id
        - title
        - slug
      properties:
        artist_id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        tonality:
          type: integer
          default: 0
        content:
          $ref: '#/components/schemas/TabContent'

    UpdateSongRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        tonality:
          type: integer
        content:
          $ref: '#/components/schemas/TabContent'