# back/internal/handler/gen/types.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for BlockKind.
const (
	Instrumental BlockKind = "instrumental"
	Lyrics       BlockKind = "lyrics"
)

// Valid indicates whether the value is a known member of the BlockKind enum.
func (e BlockKind) Valid() bool {
	switch e {
	case Instrumental:
		return true
	case Lyrics:
		return true
	default:
		return false
	}
}

// Artist defines model for Artist.
type Artist struct {
	ArtistId  openapi_types.UUID `json:"artist_id"`
	CreatedAt *time.Time         `json:"created_at,omitempty"`
	Name      string             `json:"name"`
	Slug      string             `json:"slug"`
}

// ArtistList defines model for ArtistList.
type ArtistList struct {
	Items *[]Artist `json:"items,omitempty"`
	Total *int      `json:"total,omitempty"`
}

// ArtistWithSongs defines model for ArtistWithSongs.
type ArtistWithSongs struct {
	ArtistId  openapi_types.UUID `json:"artist_id"`
	CreatedAt *time.Time         `json:"created_at,omitempty"`
	Name      string             `json:"name"`
	Slug      string             `json:"slug"`
	Songs     *[]SongListItem    `json:"songs,omitempty"`
}

// Block defines model for Block.
type Block struct {
	// Chords Only for kind=instrumental; sequence of chords
	Chords *[]string `json:"chords,omitempty"`
	Kind   BlockKind `json:"kind"`

	// Label Optional label for instrumental blocks (e.g. Intro, Proigrysh)
	Label *string `json:"label,omitempty"`

	// Segments Only for kind=lyrics; chord at segment start = switch point; text may be empty
	Segments *[]ChordSegment `json:"segments,omitempty"`

	// Tab Only for kind=instrumental; optional ASCII tab snippet
	Tab *string `json:"tab,omitempty"`
}

// BlockKind defines model for Block.Kind.
type BlockKind string

// ChordSegment defines model for ChordSegment.
type ChordSegment struct {
	Chord string `json:"chord"`

	// Text Lyrics fragment; empty string allowed (chord with no words)
	Text *string `json:"text,omitempty"`
}

// CreateArtistRequest defines model for CreateArtistRequest.
type CreateArtistRequest struct {
	Name string `json:"name"`
	Slug string `json:"slug"`
}

// CreateSongRequest defines model for CreateSongRequest.
type CreateSongRequest struct {
	ArtistId openapi_types.UUID `json:"artist_id"`
	Content  *TabContent        `json:"content,omitempty"`
	Slug     string             `json:"slug"`
	Title    string             `json:"title"`
	Tonality *int               `json:"tonality,omitempty"`
}

// Section defines model for Section.
type Section struct {
	Blocks        *[]Block  `json:"blocks,omitempty"`
	ChordSequence *[]string `json:"chord_sequence,omitempty"`
	Label         *string   `json:"label,omitempty"`
	Type          *string   `json:"type,omitempty"`
}

// Song defines model for Song.
type Song struct {
	ArtistId  *openapi_types.UUID `json:"artist_id,omitempty"`
	Content   *TabContent         `json:"content,omitempty"`
	CreatedAt *time.Time          `json:"created_at,omitempty"`
	Slug      string              `json:"slug"`
	SongId    openapi_types.UUID  `json:"song_id"`
	Title     string              `json:"title"`
	Tonality  *int                `json:"tonality,omitempty"`
	UpdatedAt *time.Time          `json:"updated_at,omitempty"`
}

// SongList defines model for SongList.
type SongList struct {
	Items *[]SongListItem `json:"items,omitempty"`
	Total *int            `json:"total,omitempty"`
}

// SongListItem defines model for SongListItem.
type SongListItem struct {
	ArtistId *openapi_types.UUID `json:"artist_id,omitempty"`
	Slug     string              `json:"slug"`
	SongId   openapi_types.UUID  `json:"song_id"`
	Title    string              `json:"title"`
	Tonality *int                `json:"tonality,omitempty"`
}

// TabContent defines model for TabContent.
type TabContent struct {
	AsciiTab *string    `json:"ascii_tab,omitempty"`
	Sections *[]Section `json:"sections,omitempty"`
}

// UpdateSongRequest defines model for UpdateSongRequest.
type UpdateSongRequest struct {
	Content  *TabContent `json:"content,omitempty"`
	Slug     *string     `json:"slug,omitempty"`
	Title    *string     `json:"title,omitempty"`
	Tonality *int        `json:"tonality,omitempty"`
}

// ListArtistsParams defines parameters for ListArtists.
type ListArtistsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// SearchParams defines parameters for Search.
type SearchParams struct {
	Q      string `form:"q" json:"q"`
	Limit  *int   `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int   `form:"offset,omitempty" json:"offset,omitempty"`
}

// ListSongsParams defines parameters for ListSongs.
type ListSongsParams struct {
	ArtistId *openapi_types.UUID `form:"artist_id,omitempty" json:"artist_id,omitempty"`
	Limit    *int                `form:"limit,omitempty" json:"limit,omitempty"`
	Offset   *int                `form:"offset,omitempty" json:"offset,omitempty"`
}

// TransposeSongParams defines parameters for TransposeSong.
type TransposeSongParams struct {
	Semitones int `form:"semitones" json:"semitones"`
}

// CreateArtistJSONRequestBody defines body for CreateArtist for application/json ContentType.
type CreateArtistJSONRequestBody = CreateArtistRequest

// CreateSongJSONRequestBody defines body for CreateSong for application/json ContentType.
type CreateSongJSONRequestBody = CreateSongRequest

// UpdateSongJSONRequestBody defines body for UpdateSong for application/json ContentType.
type UpdateSongJSONRequestBody = UpdateSongRequest